Bugfix for race condition
When logging in, it now forceably saves the session data synchronously to localStorage before the user is redirected.

This ensures that when the redirect is triggered the app state is always updated to show that you are logged in (as the call to get the session info was asyc it wasn’t previously doing this).