Refactored page logic to leverage inheritance
All pages (except HelloWorld and the clock demo) now extend the Page component. The layout logic has moved to the Layout component.

The Page component automatically populates this.props.session with the current session, when run on the server or client side. This removes the need for boiler plate to get it on every page, it’s just automatically acessible.

It would be ideal if session informaiton was a connected state variable and not a prop, so that updates to session data could automatically be propgated (and even broadcast to other open windows/tabs) but I’m going to raise that as a feature for a future update.