Refactoring sign in options to component
The sign-in buttons / form is now a component, so the view code can be displayed in a modal dialog or statically on the page. The aim is to further refactor it to dynamically show/hide options based on the providers configured.

The option to link/unlink accounts is no longer present, but will be restored in a future update on the account page (which will also show a working auth example).