Refactored to use fetch instead of XMLHttpRequest
This is code from a pull request by @emmafallancy that was merged but then reverted.

The PR had (at my suggestion!) used unfetch, but it turns out unfetch didn’t work as advertised so I sadly had to revert it.

The good news is that `isomorphic-fetch` with `credentials: 'same-origin’` set seems to work just fine now! It’s only a reduction of about 20 LOC but the code is much easier to read.