Updating to next-auth@1.7.1
NextAuth 1.7.1 includes improvements to avoid the creation of duplicate sessions (introduced in 1.6), reduced dependancies, and fixes issues with route hanlding when not passing a port in 1.7.0.

It also exposes Express in it’s response, meaning it doesn’t need to be listed as a dependancy in packages that use it to do things like add extra routes.